Fix: refactored the demo
